Create playRound() function
This is the first iteration of the function (only for the "rock" branch)
The function takes 2 inputs: playerSelection and computerSelection.
The playerSelection input is a constant ("rock") for the time being.
The computerSelection input is generated
using the getComputerChoice() function.
The function compares the two inputs, and outputs the result in the
console log based on the comparison.